Ticket #4575: fix mc-wrapper to fit changes in #4535.
As in #4535 the temporary directory seems to have switched from using a
user name based name to the process ID, and now mc-wrapper is broken. It
keeps using `whoami` in the temporary directory name: the directory
doesn't exist, the proposed output file unreachable, the returned path
is lost and the shell goes back to the initial directory.

* mc-wrapper.sh: check MC_TMPDIR environment variable. If exits, use
  it. If not, use TMPDIR with fallback to /tmp. Append process ID to
  the MC_TMPDIR value to make the directory name unique. Create mc.pwd
  file there. Restore MC_TMPDIR (w/o PID) after mc run (in case of
  mc-wrapper.sh was called via . ).
* mc-wrapper.csh: likewise.
* (mc_tmpdir): use MC_TMPDIR as is if it exists. Oherwrise use the
  same way as in wrappers.
* (main): save working directory to mc.pwd file before VFS shutdown.